added fix for removeSVGAttributes (#108)
Hi, I've found the problem with removing SVG attributes, when I'm trying to remove two or more attributes. I checked "removedAttributes" function and find that you removing the element using the index in the "for" loop, after removing the element from"svgAttrs" the next element will change position to the position of the removing item and not checked at the next iteration of the loop, so this element will be skipped. Could you please check your code.

p.s.: it's just a quick decision from my side.